Grupo2: Creacion e implementacion de una clase nueva para la Vista de Simuacion: VistaSimulacion
